+++ Ladybug/ViewModels/TableViewCellViewModel.swift
@@ -45,8 +45,51 @@ struct TableViewCellViewModel {
         self.selectAction = selectAction
     }
 
+    init(textViewString: String) {
+        self.title = textViewString
+        self.subtitle = ""
+        self.cellStyle = .default
+        self.reuseIdentifier = "textView"
+        self.selectionStyle = .none
+        self.accessoryType = .none
+        self.leadingSwipeActions = nil
+        self.trailingSwipeActions = nil
+        self.previewingViewController = nil
+        self.selectAction = {}
+    }
+
     func configure(_ cell: UITableViewCell) {
-        cell.textLabel?.text = title
+        if self.reuseIdentifier == "textView" {
+            let textView: UITextView = {
+                if let textView = cell.contentView.subviews.filter({ $0 is UITextView }).first as? UITextView {
+                    return textView
+                } else {
+                    let textView = UITextView(frame: cell.contentView.bounds)
+                    textView.dataDetectorTypes = .all
+                    textView.isScrollEnabled = false
+                    textView.font = cell.textLabel?.font
+                    textView.isEditable = false
+                    textView.isSelectable = true
+                    textView.backgroundColor = cell.backgroundColor
+                    textView.textContainerInset = .zero
+                    textView.textContainer.lineFragmentPadding = 0
+
+                    cell.contentView.addSubview(textView)
+                    textView.translatesAutoresizingMaskIntoConstraints = false
+                    textView.topAnchor.constraint(equalTo: cell.contentView.readableContentGuide.topAnchor).isActive = true
+                    textView.bottomAnchor.constraint(equalTo: cell.contentView.readableContentGuide.bottomAnchor).isActive = true
+                    textView.leftAnchor.constraint(equalTo: cell.contentView.readableContentGuide.leftAnchor).isActive = true
+                    textView.rightAnchor.constraint(equalTo: cell.contentView.readableContentGuide.rightAnchor).isActive = true
+
+                    return textView
+                }
+            }()
+
+            textView.text = title
+        } else {
+            cell.textLabel?.text = title
+        }
+
         cell.textLabel?.numberOfLines = 0
         cell.detailTextLabel?.text = subtitle
         cell.selectionStyle = selectionStyle