+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [1.2.0(11)] - 2018-07-26
 ### Added
 - [Suggestion] Add case-sensitive warning on import from Open Radar #15
 - Support radar:// links #16